move utility functions to nn_layers
